requirements.txt
Django==3.0.8
django-admin-sortable2==0.7.6
django-tinymce==3.0.2
python-dotenv==0.9.1
Pillow==7.2.0
requests==2.24.0